\name{annualised.capital.cost}
\alias{annualised.capital.cost}
\title{
Calculate annualised capital cost
}
\description{
Annualise capital cost C with discount rate r for year t
}
\usage{
annualised.capital.cost(C, r, t)
}
\arguments{
  \item{C}{
    Capital cost
}
  \item{r}{
    Discount rate
}
  \item{t}{
    Number of years
}
}
\value{
Annualised capital cost (scalar)
}